+++ native/src/main/java/io/ballerina/stdlib/grpc/Message.java
@@ -113,7 +113,7 @@ public class Message {
     private static final ArrayType stringArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_STRING);
     private static final ArrayType booleanArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_BOOLEAN);
     private static final ArrayType intArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_INT);
-    private static final ArrayType int32ArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_INT_UNSIGNED_32);
+    private static final ArrayType uInt32ArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_INT_UNSIGNED_32);
     private static final ArrayType sint32ArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_INT_SIGNED_32);
     private static final ArrayType floatArrayType = TypeCreator.createArrayType(PredefinedTypes.TYPE_FLOAT);
 
@@ -267,10 +267,15 @@ public class Message {
                             break;
                         }
                         case DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT64_VALUE:
-                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_UINT64_VALUE:
                         case DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32_VALUE:
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_UINT64_VALUE:
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_UINT32_VALUE:
                         case DescriptorProtos.FieldDescriptorProto.Type.TYPE_FIXED64_VALUE:
-                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_FIXED32_VALUE: {
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_FIXED32_VALUE:
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT64_VALUE:
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT32_VALUE:
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED64_VALUE:
+                        case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED32_VALUE: {
                             bMessage = (long) 0;
                             break;
                         }
@@ -398,7 +403,7 @@ public class Message {
                     case DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32_VALUE: {
                         if (bBMap != null) {
                             if (fieldDescriptor.isRepeated()) {
-                                BArray int32Array = ValueCreator.createArrayValue(int32ArrayType);
+                                BArray int32Array = ValueCreator.createArrayValue(sint32ArrayType);
                                 if (bBMap.containsKey(bFieldName)) {
                                     int32Array = (BArray) bBMap.get(bFieldName);
                                 } else {
@@ -429,20 +434,20 @@ public class Message {
                     case DescriptorProtos.FieldDescriptorProto.Type.TYPE_UINT32_VALUE: {
                         if (bBMap != null) {
                             if (fieldDescriptor.isRepeated()) {
-                                BArray intArray = ValueCreator.createArrayValue(intArrayType);
+                                BArray intArray = ValueCreator.createArrayValue(uInt32ArrayType);
                                 if (bBMap.containsKey(bFieldName)) {
                                     intArray = (BArray) bBMap.get(bFieldName);
                                 } else {
                                     bBMap.put(bFieldName, intArray);
                                 }
-                                intArray.add(intArray.size(), input.readUInt32());
+                                intArray.add(intArray.size(), Integer.toUnsignedLong(input.readUInt32()));
                             } else if (fieldDescriptor.getContainingOneof() != null) {
-                                updateBBMap(bBMap, fieldDescriptor, input.readUInt32());
+                                updateBBMap(bBMap, fieldDescriptor, Integer.toUnsignedLong(input.readUInt32()));
                             } else {
-                                bBMap.put(bFieldName, input.readUInt32());
+                                bBMap.put(bFieldName, Integer.toUnsignedLong(input.readUInt32()));
                             }
                         } else {
-                            bMessage = input.readUInt32();
+                            bMessage = Integer.toUnsignedLong(input.readUInt32());
                         }
                         break;
                     }
@@ -469,7 +474,7 @@ public class Message {
                     case DescriptorProtos.FieldDescriptorProto.Type.TYPE_FIXED32_VALUE: {
                         if (bBMap != null) {
                             if (fieldDescriptor.isRepeated()) {
-                                BArray int32Array = ValueCreator.createArrayValue(int32ArrayType);
+                                BArray int32Array = ValueCreator.createArrayValue(sint32ArrayType);
                                 if (bBMap.containsKey(bFieldName)) {
                                     int32Array = (BArray) bBMap.get(bFieldName);
                                 } else {
@@ -483,6 +488,86 @@ public class Message {
                             }
                         } else {
                             bMessage = input.readFixed32();
+                        }
+                        break;
+                    }
+                    case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT64_VALUE: {
+                        if (bBMap != null) {
+                            if (fieldDescriptor.isRepeated()) {
+                                BArray sInt64Array = ValueCreator.createArrayValue(intArrayType);
+                                if (bBMap.containsKey(bFieldName)) {
+                                    sInt64Array = (BArray) bBMap.get(bFieldName);
+                                } else {
+                                    bBMap.put(bFieldName, sInt64Array);
+                                }
+                                sInt64Array.add(sInt64Array.size(), input.readSInt64());
+                            } else if (fieldDescriptor.getContainingOneof() != null) {
+                                updateBBMap(bBMap, fieldDescriptor, input.readSInt64());
+                            } else {
+                                bBMap.put(bFieldName, input.readSInt64());
+                            }
+                        } else {
+                            bMessage = input.readSInt64();
+                        }
+                        break;
+                    }
+                    case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT32_VALUE: {
+                        if (bBMap != null) {
+                            if (fieldDescriptor.isRepeated()) {
+                                BArray sInt32Array = ValueCreator.createArrayValue(sint32ArrayType);
+                                if (bBMap.containsKey(bFieldName)) {
+                                    sInt32Array = (BArray) bBMap.get(bFieldName);
+                                } else {
+                                    bBMap.put(bFieldName, sInt32Array);
+                                }
+                                sInt32Array.add(sInt32Array.size(), input.readSInt32());
+                            } else if (fieldDescriptor.getContainingOneof() != null) {
+                                updateBBMap(bBMap, fieldDescriptor, input.readSInt32());
+                            } else {
+                                bBMap.put(bFieldName, input.readSInt32());
+                            }
+                        } else {
+                            bMessage = input.readSInt32();
+                        }
+                        break;
+                    }
+                    case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED64_VALUE: {
+                        if (bBMap != null) {
+                            if (fieldDescriptor.isRepeated()) {
+                                BArray sFixed64Array = ValueCreator.createArrayValue(intArrayType);
+                                if (bBMap.containsKey(bFieldName)) {
+                                    sFixed64Array = (BArray) bBMap.get(bFieldName);
+                                } else {
+                                    bBMap.put(bFieldName, sFixed64Array);
+                                }
+                                sFixed64Array.add(sFixed64Array.size(), input.readSFixed64());
+                            } else if (fieldDescriptor.getContainingOneof() != null) {
+                                updateBBMap(bBMap, fieldDescriptor, input.readSFixed64());
+                            } else {
+                                bBMap.put(bFieldName, input.readSFixed64());
+                            }
+                        } else {
+                            bMessage = input.readSFixed64();
+                        }
+                        break;
+                    }
+                    case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED32_VALUE: {
+                        if (bBMap != null) {
+                            if (fieldDescriptor.isRepeated()) {
+                                BArray sFixed32Array = ValueCreator.createArrayValue(sint32ArrayType);
+                                if (bBMap.containsKey(bFieldName)) {
+                                    sFixed32Array = (BArray) bBMap.get(bFieldName);
+                                } else {
+                                    bBMap.put(bFieldName, sFixed32Array);
+                                }
+                                sFixed32Array.add(sFixed32Array.size(), input.readSFixed32());
+                            } else if (fieldDescriptor.getContainingOneof() != null) {
+                                updateBBMap(bBMap, fieldDescriptor, input.readSFixed32());
+                            } else {
+                                bBMap.put(bFieldName, input.readSFixed32());
+                            }
+                        } else {
+                            bMessage = input.readSFixed32();
                         }
                         break;
                     }
@@ -928,13 +1013,16 @@ public class Message {
                         if (bValue instanceof BArray) {
                             BArray valueArray = (BArray) bValue;
                             for (int i = 0; i < valueArray.size(); i++) {
-                                output.writeUInt32(fieldDescriptor.getNumber(), (int) valueArray.getInt(i));
+                                output.writeUInt32(fieldDescriptor.getNumber(),
+                                        Integer.parseUnsignedInt(String.valueOf(valueArray.getInt(i))));
                             }
                         } else {
-                            output.writeUInt32(fieldDescriptor.getNumber(), (int) bValue);
+                            output.writeUInt32(fieldDescriptor.getNumber(),
+                                    Integer.parseUnsignedInt(String.valueOf(bValue)));
                         }
                     } else if (bMessage instanceof Long) {
-                        output.writeUInt32(fieldDescriptor.getNumber(), (int) bMessage);
+                        output.writeUInt32(fieldDescriptor.getNumber(),
+                                Integer.parseUnsignedInt(String.valueOf(bMessage)));
                     }
                     break;
                 }
@@ -960,14 +1048,77 @@ public class Message {
                         if (bValue instanceof BArray) {
                             BArray valueArray = (BArray) bValue;
                             for (int i = 0; i < valueArray.size(); i++) {
-                                output.writeFixed32(fieldDescriptor.getNumber(),
-                                                    getIntValue(valueArray.getInt(i)));
+                                output.writeFixed32(fieldDescriptor.getNumber(), getIntValue(valueArray.getInt(i)));
                             }
                         } else {
                             output.writeFixed32(fieldDescriptor.getNumber(), getIntValue(bValue));
                         }
                     } else if (bMessage instanceof Long) {
                         output.writeFixed32(fieldDescriptor.getNumber(), getIntValue(bMessage));
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT64_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                output.writeSInt64(fieldDescriptor.getNumber(), valueArray.getInt(i));
+                            }
+                        } else {
+                            output.writeSInt64(fieldDescriptor.getNumber(), (long) bValue);
+                        }
+                    } else if (bMessage instanceof Long) {
+                        output.writeSInt64(fieldDescriptor.getNumber(), (long) bMessage);
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT32_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                output.writeSInt32(fieldDescriptor.getNumber(), getIntValue(valueArray.getInt(i)));
+                            }
+                        } else {
+                            output.writeSInt32(fieldDescriptor.getNumber(), getIntValue(bValue));
+                        }
+                    } else if (bMessage instanceof Long) {
+                        output.writeSInt32(fieldDescriptor.getNumber(), getIntValue(bMessage));
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED64_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                output.writeSFixed64(fieldDescriptor.getNumber(), valueArray.getInt(i));
+                            }
+                        } else {
+                            output.writeSFixed64(fieldDescriptor.getNumber(), (long) bValue);
+                        }
+                    } else if (bMessage instanceof Long) {
+                        output.writeSFixed64(fieldDescriptor.getNumber(), (long) bMessage);
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED32_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                output.writeSFixed32(fieldDescriptor.getNumber(), getIntValue(valueArray.getInt(i)));
+                            }
+                        } else {
+                            output.writeSFixed32(fieldDescriptor.getNumber(), getIntValue(bValue));
+                        }
+                    } else if (bMessage instanceof Long) {
+                        output.writeSFixed32(fieldDescriptor.getNumber(), getIntValue(bMessage));
                     }
                     break;
                 }
@@ -1276,15 +1427,16 @@ public class Message {
                             BArray valueArray = (BArray) bValue;
                             for (int i = 0; i < valueArray.size(); i++) {
                                 size += com.google.protobuf.CodedOutputStream.computeUInt32Size(
-                                        fieldDescriptor.getNumber(), (int) valueArray.getInt(i));
+                                        fieldDescriptor.getNumber(),
+                                        Integer.parseUnsignedInt(String.valueOf(valueArray.getInt(i))));
                             }
                         } else {
                             size += com.google.protobuf.CodedOutputStream.computeUInt32Size(
-                                    fieldDescriptor.getNumber(), (int) bValue);
+                                    fieldDescriptor.getNumber(), Integer.parseUnsignedInt(String.valueOf(bValue)));
                         }
                     } else if (bMessage instanceof Long) {
                         size += com.google.protobuf.CodedOutputStream.computeUInt32Size(fieldDescriptor
-                                .getNumber(), (int) bMessage);
+                                .getNumber(), Integer.parseUnsignedInt(String.valueOf(bMessage)));
                     }
                     break;
                 }
@@ -1322,6 +1474,82 @@ public class Message {
                         }
                     } else if (bMessage instanceof Long) {
                         size += com.google.protobuf.CodedOutputStream.computeFixed32Size(fieldDescriptor
+                                .getNumber(), getIntValue(bMessage));
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT64_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                size += com.google.protobuf.CodedOutputStream.computeSInt64Size(
+                                        fieldDescriptor.getNumber(), valueArray.getInt(i));
+                            }
+                        } else {
+                            size += com.google.protobuf.CodedOutputStream.computeSInt64Size(
+                                    fieldDescriptor.getNumber(), (long) bValue);
+                        }
+                    } else if (bMessage instanceof Long) {
+                        size += com.google.protobuf.CodedOutputStream.computeSInt64Size(fieldDescriptor
+                                .getNumber(), (long) bMessage);
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SINT32_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                size += com.google.protobuf.CodedOutputStream.computeSInt32Size(
+                                        fieldDescriptor.getNumber(), getIntValue(valueArray.getInt(i)));
+                            }
+                        } else {
+                            size += com.google.protobuf.CodedOutputStream.computeSInt32Size(
+                                    fieldDescriptor.getNumber(), getIntValue(bValue));
+                        }
+                    } else if (bMessage instanceof Long) {
+                        size += com.google.protobuf.CodedOutputStream.computeSInt32Size(fieldDescriptor
+                                .getNumber(), getIntValue(bMessage));
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED64_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                size += com.google.protobuf.CodedOutputStream.computeSFixed64Size(
+                                        fieldDescriptor.getNumber(), valueArray.getInt(i));
+                            }
+                        } else {
+                            size += com.google.protobuf.CodedOutputStream.computeSFixed64Size(
+                                    fieldDescriptor.getNumber(), (long) bValue);
+                        }
+                    } else if (bMessage instanceof Long) {
+                        size += com.google.protobuf.CodedOutputStream.computeSFixed64Size(fieldDescriptor
+                                .getNumber(), (long) bMessage);
+                    }
+                    break;
+                }
+                case DescriptorProtos.FieldDescriptorProto.Type.TYPE_SFIXED32_VALUE: {
+                    if (bBMap != null && bBMap.containsKey(bFieldName)) {
+                        Object bValue = bBMap.get(bFieldName);
+                        if (bValue instanceof BArray) {
+                            BArray valueArray = (BArray) bValue;
+                            for (int i = 0; i < valueArray.size(); i++) {
+                                size += com.google.protobuf.CodedOutputStream.computeSFixed32Size(
+                                        fieldDescriptor.getNumber(), getIntValue(valueArray.getInt(i)));
+                            }
+                        } else {
+                            size += com.google.protobuf.CodedOutputStream.computeSFixed32Size(
+                                    fieldDescriptor.getNumber(), getIntValue(bValue));
+                        }
+                    } else if (bMessage instanceof Long) {
+                        size += com.google.protobuf.CodedOutputStream.computeSFixed32Size(fieldDescriptor
                                 .getNumber(), getIntValue(bMessage));
                     }
                     break;