+++ changelog.md
@@ -1,5 +1,5 @@
 # Change Log
-This file contains all the notable changes done to the Ballerina time package through the releases.
+This file contains all the notable changes done to the Ballerina gRPC package through the releases.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - [Updated Protocol Buffers version to 3.21.7](https://github.com/ballerina-platform/ballerina-standard-library/issues/3478)
 - [API Docs Updated](https://github.com/ballerina-platform/ballerina-standard-library/issues/3463)
+
+### Fixed
+- [Missing support for different int types](https://github.com/ballerina-platform/ballerina-standard-library/issues/3558)
 
 ### [1.4.1] - 2022-10-06
 ### Fixed